# improve toFixed/toString performance